requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
yfinance==0.2.43
pandas==2.2.2
google-generativeai==0.8.0
httpx==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
numpy==1.26.4
